requirements.txt
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.8.1
alibabacloud-credentials==0.3.0
alibabacloud-dysmsapi20170525==2.0.22
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.1
alibabacloud-openapi-util==0.2.0
alibabacloud-tea==0.3.0
alibabacloud-tea-openapi==0.3.6
alibabacloud-tea-util==0.3.8
alibabacloud-tea-xml==0.0.2
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.0
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
asyncmy==0.2.7
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
crcmod==1.7
cryptography==38.0.3
dnspython==2.2.1
ecdsa==0.18.0
et-xmlfile==1.1.0
fastapi==0.95.0
frozenlist==1.3.3
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
jmespath==0.10.0
loguru==0.6.0
Mako==1.2.4
MarkupSafe==2.1.1
motor==3.1.1
multidict==6.0.2
openpyxl==3.0.10
oss2==2.16.0
passlib==1.7.4
Pillow==9.3.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.3
PyMySQL==1.0.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
SQLAlchemy-Utils==0.38.3
SSIM-PIL==1.0.14
starlette==0.26.1
typer==0.7.0
typing-extensions==4.4.0
ua-parser==0.16.1
urllib3==1.26.12
user-agents==2.2.0
uvicorn==0.21.0
watchgod==0.8.2
websockets==10.4
win32-setctime==1.1.0
XlsxWriter==3.0.3
yarl==1.8.1
zipp==3.10.0
zope.event==4.5.0
zope.interface==5.5.2
orjson==3.8.2